fix (deploy): CSR
