+++ .github/workflows/cd.yml
@@ -41,7 +41,7 @@ jobs:
       - name: Upload Build
         uses: actions/upload-pages-artifact@v4
         with:
-          path: "dist/client/"
+          path: "dist/"
 
       - name: Deploy to GitHub Pages
         uses: actions/deploy-pages@v4
