+++ index.html
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta http-equiv="Content-Type" content="text/html;charset=UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+
+    <title>SASE LSU</title>
+    <meta name="description" content="Welcome to the SASE@LSU website!">
+    <meta name="keywords" content="sase, lsu">
+
+    <meta property="og:title" content="SASE LSU" />
+    <meta property="og:description" content="Welcome to the SASE@LSU website!" />
+    <meta property="og:url" content="www.saselsu.org" />
+
+    <meta name="theme-color" content="#000000" />
+
+    <link rel="icon" href="/favicon.ico">
+    <link rel="apple-touch-icon" href="/logo192.png">
+    <link rel="manifest" href="./manifest.json">
+</head>
+<body>
+    <noscript>You need to enable JavaScript to run this app.</noscript>
+    <div id="root"></div>
+
+    <script type="module" src="./src/index.tsx"></script>
+</body>
+</html>
