+++ package.json
@@ -5,8 +5,8 @@
   "type": "module",
   "homepage": "https://www.saselsu.org",
   "scripts": {
-    "build": " bun --bun react-router build",
-    "dev": "bun --bun react-router dev",
+    "build": " bun --bun vite build",
+    "dev": "bun --bun vite dev",
     "preview": "bun --bun vite preview",
     "lint": "eslint . --fix",
     "lint:ci": "eslint ."
@@ -36,6 +36,7 @@
     "@types/react": "^19.1.13",
     "@types/react-dom": "^19.1.9",
     "@types/three": "^0.180.0",
+    "@vitejs/plugin-react": "^5.0.3",
     "eslint": "^9.36.0",
     "eslint-plugin-jsx-a11y": "^6.10.2",
     "eslint-plugin-react": "^7.37.5",
