--- react-router.config.ts
@@ -1,7 +1,0 @@
-import type { Config } from "@react-router/dev/config";
-
-export default {
-    buildDirectory: "dist",
-    appDirectory: "src",
-    ssr: false
-} satisfies Config;
