+++ src/index.tsx
@@ -1,0 +1,9 @@
+import { StrictMode } from "react";
+import { createRoot } from "react-dom/client";
+import App from "./App";
+
+createRoot(document.querySelector("#root")!).render(
+    <StrictMode>
+        <App />
+    </StrictMode>
+);
