+++ src/vite-env.d.ts
@@ -1,0 +1,8 @@
+/// <reference types="vite/client" />
+
+/**
+ * TSC fails to resolve images.
+ */
+declare module "*.png";
+declare module "*.jpg";
+declare module "*.jpeg";
