+++ tsconfig.json
@@ -6,12 +6,6 @@
 
     "moduleResolution": "Bundler",
     "jsx": "react-jsx",
-    "rootDirs": [".", "./.react-router/types"],
-    "baseUrl": ".",
-    "paths": {
-      "~/*": ["./src/*"]
-    },
-    "types": ["node", "vite/client"],
 
     "strict": true,
     "strictNullChecks": true,
@@ -33,9 +27,7 @@
     "src/**/*.tsx",
     "src/**/*.jsx",
     "src/**/*.js",
-    "vite.config.ts",
-    "react-router.config.ts",
-    ".react-router/types/**/*"
+    "vite.config.ts"
   ],
   "exclude": [
     "node_modules",
